# Use Z3's API