--- a/silicon.sh
+++ b/silicon.sh
@@ -4,8 +4,6 @@
 
 set -e
 
-set -e
-
 BASEDIR="$(realpath `dirname $0`)"
 
 CP_FILE="$BASEDIR/silicon_classpath.txt"
@@ -23,4 +21,4 @@ if [ ! -f "$LIB_Z3" ]; then
      curl 'http://www.sosy-lab.org/ivy/org.sosy_lab/javasmt-solver-z3/libz3-4.8.7.so' -o "$LIB_Z3"
 fi
 
-java -Xss30M -Djava.library.path="$(pwd)/" -Dlogback.configurationFile="$BASEDIR/src/main/resources/logback.xml" -cp "`cat $CP_FILE`" viper.silicon.SiliconRunner $@
+java -Xss30M -Djava.library.path="$BASEDIR" -Dlogback.configurationFile="$BASEDIR/src/main/resources/logback.xml" -cp "`cat $CP_FILE`" viper.silicon.SiliconRunner $@