Do not merge. This PR is a proof of concept for using Z3's API from Silicon. On a few Viper examples that I tried it seems to work fine, with up to a 2x slowdown (e.g. 58s -> 2m). This is expected as Silicon is still pretty-printing all the commands as before, so there is room for optimizations.
* The `Z3_EXE` variable is no longer used.
* The code has not been optimized, and is e.g. quadratic in the number of sort and function definitions.
* Various hacks are used to still support clients who want to directly emit SMT-LIB2 commands (e.g. in a case we pretty print a term, match it with a regex, parse it with a SMT-LIB2 parser, use it to generate another SMT-LIB2 command, and finally pass the command to Z3, which will use its own parser).
* The PR overwrites the existing `Z3ProverStdIO` just to have a clear diff of the modifications.

What this PR doesn't do:
* [ ] Move the modifications to a new `Z3ProverLib` class, adding a command line option to choose between it and the usual `Z3ProverStdIO`.
* [ ] Rewrite clients of `Z3ProverStdIO` to rely less on `emit(String)` and more on a new `emitTerm(Term)` method.
* [ ] Implement a conversion from `Term` to Z3's AST without going through the SMT-LIB2 format. This will allow to remove hacks and conversions of `Term` types to `String`.
* [ ] Compile Z3 enabling the Java API. This means that the Z3 artifacts produced by Jenkins should include e.g. the `libz3.so`, `libz3java.so` and `z3.jar` files.