@fpoli 2020-11-10
Interestingly, on small programs like those in the test suite the verification time already improved.

Benchmark of `examples/binary-search` before this PR:
```
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-array.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:     25 msec (Parsing),     14 msec (Semantic Analysis),      8 msec (Translation),     18 msec (Consistency Check),   1.63 sec  (Verification),   1.69 sec  (Overall). 
[info]   + [Benchmark] Stddevs:             8 msec (Parsing),      4 msec (Semantic Analysis),      2 msec (Translation),      6 msec (Consistency Check),    101 msec (Verification),    119 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:        32 %    (Parsing),     28 %    (Semantic Analysis),     25 %    (Translation),     33 %    (Consistency Check),      6 %    (Verification),      7 %    (Overall). 
[info]   + [Benchmark] Best run:           17 msec (Parsing),      9 msec (Semantic Analysis),      6 msec (Translation),     15 msec (Consistency Check),   1.54 sec  (Verification),   1.58 sec  (Overall). 
[info]   + [Benchmark] Median run:         27 msec (Parsing),     14 msec (Semantic Analysis),      7 msec (Translation),     18 msec (Consistency Check),   1.61 sec  (Verification),   1.67 sec  (Overall). 
[info]   + [Benchmark] Worst run:          46 msec (Parsing),     24 msec (Semantic Analysis),     11 msec (Translation),     35 msec (Consistency Check),   1.85 sec  (Verification),   1.96 sec  (Overall). 
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-seq.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:      5 msec (Parsing),      6 msec (Semantic Analysis),      3 msec (Translation),      8 msec (Consistency Check),   1.20 sec  (Verification),   1.22 sec  (Overall). 
[info]   + [Benchmark] Stddevs:             1 msec (Parsing),      1 msec (Semantic Analysis),      0 msec (Translation),      1 msec (Consistency Check),     60 msec (Verification),     58 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:        20 %    (Parsing),     16 %    (Semantic Analysis),      0 %    (Translation),     12 %    (Consistency Check),      5 %    (Verification),      4 %    (Overall). 
[info]   + [Benchmark] Best run:            6 msec (Parsing),      6 msec (Semantic Analysis),      3 msec (Translation),     12 msec (Consistency Check),   1.12 sec  (Verification),   1.15 sec  (Overall). 
[info]   + [Benchmark] Median run:          5 msec (Parsing),      5 msec (Semantic Analysis),      3 msec (Translation),      8 msec (Consistency Check),   1.18 sec  (Verification),   1.20 sec  (Overall). 
[info]   + [Benchmark] Worst run:           5 msec (Parsing),      5 msec (Semantic Analysis),      3 msec (Translation),      7 msec (Consistency Check),   1.31 sec  (Verification),   1.33 sec  (Overall). 
```

After this PR:
```
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-array.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:     25 msec (Parsing),     13 msec (Semantic Analysis),      6 msec (Translation),     18 msec (Consistency Check),   1.20 sec  (Verification),   1.26 sec  (Overall). 
[info]   + [Benchmark] Stddevs:            10 msec (Parsing),      4 msec (Semantic Analysis),      1 msec (Translation),      2 msec (Consistency Check),    130 msec (Verification),    146 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:        40 %    (Parsing),     30 %    (Semantic Analysis),     16 %    (Translation),     11 %    (Consistency Check),     10 %    (Verification),     11 %    (Overall). 
[info]   + [Benchmark] Best run:           17 msec (Parsing),     11 msec (Semantic Analysis),      7 msec (Translation),     17 msec (Consistency Check),   1.06 sec  (Verification),   1.11 sec  (Overall). 
[info]   + [Benchmark] Median run:         24 msec (Parsing),     11 msec (Semantic Analysis),      7 msec (Translation),     17 msec (Consistency Check),   1.19 sec  (Verification),   1.25 sec  (Overall). 
[info]   + [Benchmark] Worst run:          50 msec (Parsing),     21 msec (Semantic Analysis),      8 msec (Translation),     22 msec (Consistency Check),   1.50 sec  (Verification),   1.60 sec  (Overall). 
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-seq.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:      6 msec (Parsing),      8 msec (Semantic Analysis),      4 msec (Translation),     10 msec (Consistency Check),    932 msec (Verification),    962 msec (Overall). 
[info]   + [Benchmark] Stddevs:             1 msec (Parsing),      2 msec (Semantic Analysis),      1 msec (Translation),      0 msec (Consistency Check),     11 msec (Verification),     14 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:        16 %    (Parsing),     25 %    (Semantic Analysis),     25 %    (Translation),      0 %    (Consistency Check),      1 %    (Verification),      1 %    (Overall). 
[info]   + [Benchmark] Best run:            6 msec (Parsing),      7 msec (Semantic Analysis),      4 msec (Translation),     10 msec (Consistency Check),    916 msec (Verification),    943 msec (Overall). 
[info]   + [Benchmark] Median run:          6 msec (Parsing),      7 msec (Semantic Analysis),      4 msec (Translation),     10 msec (Consistency Check),    937 msec (Verification),    964 msec (Overall). 
[info]   + [Benchmark] Worst run:           7 msec (Parsing),      8 msec (Semantic Analysis),      4 msec (Translation),     10 msec (Consistency Check),    957 msec (Verification),    986 msec (Overall). 
```