@mschwerhoff 2020-11-11
I'm curious to see if performance can be improved significantly without replacing Silicon's terms with their Z3 API counterparts everywhere. Let me know if you'd like to discuss ideas or if you have any questions.