@fpoli 2020-11-11
> Interestingly, on small programs like those in the test suite the verification time already improved a bit.

No longer true after fd7e0d669d76279573ed5e2350a16431113c912d, which fixes a Silicon slowdown.

Before the PR:
```
[info] PortableSiliconTests:
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-array.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:     28 msec (Parsing),     15 msec (Semantic Analysis),      8 msec (Translation),     19 msec (Consistency Check),    922 msec (Verification),    993 msec (Overall). 
[info]   + [Benchmark] Stddevs:            13 msec (Parsing),      5 msec (Semantic Analysis),      2 msec (Translation),      4 msec (Consistency Check),    115 msec (Verification),    138 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:        46 %    (Parsing),     33 %    (Semantic Analysis),     25 %    (Translation),     21 %    (Consistency Check),     12 %    (Verification),     13 %    (Overall). 
[info]   + [Benchmark] Best run:           21 msec (Parsing),     12 msec (Semantic Analysis),      7 msec (Translation),     16 msec (Consistency Check),    807 msec (Verification),    863 msec (Overall). 
[info]   + [Benchmark] Median run:         24 msec (Parsing),     12 msec (Semantic Analysis),      7 msec (Translation),     17 msec (Consistency Check),    920 msec (Verification),    980 msec (Overall). 
[info]   + [Benchmark] Worst run:          62 msec (Parsing),     29 msec (Semantic Analysis),     12 msec (Translation),     29 msec (Consistency Check),   1.17 sec  (Verification),   1.30 sec  (Overall). 
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-seq.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:      5 msec (Parsing),      6 msec (Semantic Analysis),      3 msec (Translation),      8 msec (Consistency Check),    569 msec (Verification),    592 msec (Overall). 
[info]   + [Benchmark] Stddevs:             0 msec (Parsing),      0 msec (Semantic Analysis),      0 msec (Translation),      1 msec (Consistency Check),     24 msec (Verification),     25 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:         0 %    (Parsing),      0 %    (Semantic Analysis),      0 %    (Translation),     12 %    (Consistency Check),      4 %    (Verification),      4 %    (Overall). 
[info]   + [Benchmark] Best run:            5 msec (Parsing),      6 msec (Semantic Analysis),      3 msec (Translation),      8 msec (Consistency Check),    538 msec (Verification),    560 msec (Overall). 
[info]   + [Benchmark] Median run:          6 msec (Parsing),      5 msec (Semantic Analysis),      4 msec (Translation),      8 msec (Consistency Check),    563 msec (Verification),    586 msec (Overall). 
[info]   + [Benchmark] Worst run:           5 msec (Parsing),      6 msec (Semantic Analysis),      4 msec (Translation),      8 msec (Consistency Check),    618 msec (Verification),    641 msec (Overall). 
```

After the PR:
```
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-array.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:     24 msec (Parsing),     13 msec (Semantic Analysis),      7 msec (Translation),     17 msec (Consistency Check),   1.17 sec  (Verification),   1.24 sec  (Overall). 
[info]   + [Benchmark] Stddevs:            12 msec (Parsing),      4 msec (Semantic Analysis),      2 msec (Translation),      4 msec (Consistency Check),    107 msec (Verification),    128 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:        50 %    (Parsing),     30 %    (Semantic Analysis),     28 %    (Translation),     23 %    (Consistency Check),      9 %    (Verification),     10 %    (Overall). 
[info]   + [Benchmark] Best run:           18 msec (Parsing),     10 msec (Semantic Analysis),      6 msec (Translation),     15 msec (Consistency Check),   1.06 sec  (Verification),   1.11 sec  (Overall). 
[info]   + [Benchmark] Median run:         19 msec (Parsing),     11 msec (Semantic Analysis),      6 msec (Translation),     15 msec (Consistency Check),   1.17 sec  (Verification),   1.22 sec  (Overall). 
[info]   + [Benchmark] Worst run:          55 msec (Parsing),     24 msec (Semantic Analysis),     12 msec (Translation),     21 msec (Consistency Check),   1.40 sec  (Verification),   1.51 sec  (Overall). 
[info] - ../silver/src/test/resources/examples/binary-search/binary-search-seq.vpr [Silicon-Statistics-Silver]
[info]   + [Benchmark] Trimmed 2 marginal runs. 
[info]   + [Benchmark] Mean /  8 runs:      5 msec (Parsing),      5 msec (Semantic Analysis),      3 msec (Translation),      7 msec (Consistency Check),    729 msec (Verification),    750 msec (Overall). 
[info]   + [Benchmark] Stddevs:             0 msec (Parsing),      0 msec (Semantic Analysis),      0 msec (Translation),      0 msec (Consistency Check),     14 msec (Verification),     14 msec (Overall). 
[info]   + [Benchmark] Rel. stddev:         0 %    (Parsing),      0 %    (Semantic Analysis),      0 %    (Translation),      0 %    (Consistency Check),      1 %    (Verification),      1 %    (Overall). 
[info]   + [Benchmark] Best run:            5 msec (Parsing),      5 msec (Semantic Analysis),      3 msec (Translation),      9 msec (Consistency Check),    708 msec (Verification),    730 msec (Overall). 
[info]   + [Benchmark] Median run:          6 msec (Parsing),      7 msec (Semantic Analysis),      4 msec (Translation),      7 msec (Consistency Check),    731 msec (Verification),    755 msec (Overall). 
[info]   + [Benchmark] Worst run:           5 msec (Parsing),      6 msec (Semantic Analysis),      3 msec (Translation),      7 msec (Consistency Check),    753 msec (Verification),    774 msec (Overall). 
```

> I'm curious to see if performance can be improved significantly without replacing Silicon's terms with their Z3 API counterparts everywhere. Let me know if you'd like to discuss ideas or if you have any questions.

I pushed a commit as that's what I was trying. The SMTLIB2 parser in Z3's API still requires function and sort declarations to be declared using the API. So, the `emit(..)` method needs to use the ugly regexp and SMTLIB2 parsing to intercept declarations that come from the prelude. Avoiding this and changing the format of the prelude would require substantial work, so I'm not trying it.

As it is now, we are running the same code as before to convert terms to string and then to parse them in Z3. So, it's not surprising that there is no performance benefit. It might be that constructing everything via the Z3 API is faster, but also trying that requires substantial work.